Create bryant_perceptron_test
Edited sign function; tested perceptron. Didn't edit Perceptron class.

The next thing we might look at is calculating the tolerance. The problem right now is that, as written, it calculates if a change needs to be made every time it updates. So, even if a couple of changes sum to over the tolerance level, it's not going to update if any individual one isn't above the tolerance level. We could just lower the tolerance level, but it might have to get really small when dealing with large data.